Changes to make semargl compatible with java modularization (removed packages defined in several maven modules).
