+++ core/src/main/java/module-info.java
@@ -1,0 +1,9 @@
+module semargl.core {
+    requires  java.xml;
+    exports org.semarglproject.rdf.core;
+    exports org.semarglproject.sink;
+    exports org.semarglproject.ri;
+    exports org.semarglproject.source;
+    exports org.semarglproject.vocab.core;
+    exports org.semarglproject.xml;
+}
