--- core/src/main/java/org/semarglproject/rdf/ParseException.java
@@ -1,33 +1,0 @@
-/**
- * Copyright 2012-2013 the Semargl contributors. See AUTHORS for more details.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.semarglproject.rdf;
-
-public class ParseException extends Exception {
-
-    private static final long serialVersionUID = 2088926094965976520L;
-
-    public ParseException(String string) {
-        super(string);
-    }
-
-    public ParseException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public ParseException(Throwable cause) {
-        super(cause);
-    }
-}
