+++ core/src/main/java/org/semarglproject/sink/CharOutputSink.java
@@ -15,7 +15,7 @@
  */
 package org.semarglproject.sink;
 
-import org.semarglproject.rdf.ParseException;
+import org.semarglproject.rdf.core.ParseException;
 
 import java.io.Closeable;
 import java.io.File;
