+++ core/src/main/java/org/semarglproject/source/CharSource.java
@@ -15,7 +15,7 @@
  */
 package org.semarglproject.source;
 
-import org.semarglproject.rdf.ParseException;
+import org.semarglproject.rdf.core.ParseException;
 import org.semarglproject.sink.CharSink;
 
 import java.io.BufferedReader;
