+++ jsonld/src/main/java/module-info.java
@@ -1,0 +1,4 @@
+module semargl.jsonld {
+    requires semargl.core;
+    requires semargl.rdf;
+}
