+++ jsonld/src/main/java/org/semarglproject/jsonld/JsonLdContentHandler.java
@@ -18,8 +18,8 @@ package org.semarglproject.jsonld;
 
 import org.semarglproject.ri.MalformedIriException;
 import org.semarglproject.sink.QuadSink;
-import org.semarglproject.vocab.JsonLd;
-import org.semarglproject.vocab.RDF;
+import org.semarglproject.vocab.jsonld.JsonLd;
+import org.semarglproject.vocab.core.RDF;
 import org.semarglproject.vocab.XSD;
 
 import java.util.Deque;
