+++ jsonld/src/main/java/org/semarglproject/jsonld/JsonLdParser.java
@@ -16,8 +16,8 @@
 
 package org.semarglproject.jsonld;
 
-import org.semarglproject.rdf.ParseException;
-import org.semarglproject.rdf.ProcessorGraphHandler;
+import org.semarglproject.rdf.core.ParseException;
+import org.semarglproject.rdf.core.ProcessorGraphHandler;
 import org.semarglproject.sink.CharSink;
 import org.semarglproject.sink.Pipe;
 import org.semarglproject.sink.QuadSink;
