+++ rdf/src/main/java/module-info.java
@@ -1,0 +1,6 @@
+module semargl.rdf {
+    requires semargl.core;
+    requires java.xml;
+    exports org.semarglproject.rdf;
+    exports org.semarglproject.vocab;
+}
