+++ rdf/src/main/java/org/semarglproject/rdf/NQuadsSerializer.java
@@ -15,6 +15,7 @@
  */
 package org.semarglproject.rdf;
 
+import org.semarglproject.rdf.core.ParseException;
 import org.semarglproject.sink.CharSink;
 import org.semarglproject.sink.QuadSink;
