+++ rdfa/src/main/java/module-info.java
@@ -1,0 +1,7 @@
+module semargl.rdfa {
+    requires semargl.core;
+    requires semargl.rdf;
+    requires java.xml;
+    requires java.xml.bind;
+    exports org.semarglproject.rdf.rdfa;
+}
